Reorganise between util functions and format-importer methods
