+++ util.ts
@@ -58,40 +58,10 @@ export function genUid(length: number): string {
 }
 
 /**
- * Adds a single tag to the tag property in frontmatter and santises it.
- * Must pass in app.fileManager.
+ * Takes a string and returns in lowercase with the first letter capitalised.
  */
-export async function addTagToFrontmatter(tag: string, fileRef: TFile, fileManager: FileManager) {
-	const sanitizedTag = sanitizeTag(tag);
-	await fileManager.processFrontMatter(fileRef, (frontmatter: any) => {
-		if(!frontmatter['tags']) {
-			frontmatter['tags'] = [sanitizedTag];
-		} else {
-			if (!Array.isArray(frontmatter['tags'])) {
-				frontmatter['tags'] = frontmatter['tags'].split(' ');
-			}
-			frontmatter['tags'].push(sanitizedTag);
-		}
-	});
-}
-
-/**
- * Adds an alias to the note's frontmatter.
- * Only linebreak sanitization is performed in this function.
- * Must pass in app.fileManager.
-*/
-export async function addAliasToFrontmatter(alias: string, fileRef: TFile, fileManager: FileManager) {
-	const sanitizedAlias = alias.split('\n').join(', ');
-	await fileManager.processFrontMatter(fileRef, (frontmatter: any) => {      
-		if(!frontmatter['aliases']) {
-			frontmatter['aliases'] = [sanitizedAlias];
-		} else {
-			if (!Array.isArray(frontmatter['aliases'])) {
-				frontmatter['aliases'] = frontmatter['aliases'].split(' ');
-			}
-			frontmatter['aliases'].push(sanitizedAlias);
-		}
-	});
+export function toSentenceCase(str: string) {
+    return str.charAt(0).toUpperCase() + str.slice(1).toLowerCase();
 }
 
 export class PromiseExecutor {
